feat(edit-dashboard-json): [MTM-61842] Update dashboard refactor
